# Add option to build in the same directory, to better use caches